+++ src/vfs/mod.rs
@@ -1,5 +1,6 @@
 mod dir_fs;
 mod dir_entry;
+mod map_fs;
 
 pub use dir_fs::DirFS;
 pub use dir_entry::{DirEntry, DirEntryType};